fix: clippy lints

+++ ginepro/examples/client.rs
@@ -5,7 +5,7 @@ use shared_proto::pb::EchoRequest;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let channel = LoadBalancedChannelBuilder::new_with_service(("localhost", 5000 as u16))
+    let channel = LoadBalancedChannelBuilder::new_with_service(("localhost", 5000_u16))
         .await?
         .dns_probe_interval(std::time::Duration::from_secs(5))
         .channel();
