+++ ginepro/examples/client_tls.rs
@@ -13,7 +13,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let tls = ClientTlsConfig::new().ca_certificate(ca);
 
-    let channel = LoadBalancedChannelBuilder::new_with_service(("localhost", 5000 as u16))
+    let channel = LoadBalancedChannelBuilder::new_with_service(("localhost", 5000_u16))
         .await?
         .with_tls(tls)
         .dns_probe_interval(std::time::Duration::from_secs(5))
